+++ VWOSdk.Tests/VWOTests.cs
@@ -100,7 +100,7 @@ namespace VWOSdk.Tests
             var mockSettingProcessor = Mock.GetSettingsProcessor();
             VWO.Configure(mockSettingProcessor.Object);
 
-            var vwoClient = VWO.CreateInstance(validSettings);
+            var vwoClient = VWO.Launch(validSettings);
             Assert.NotNull(vwoClient);
             Assert.IsType<VWO>(vwoClient);
 
@@ -118,7 +118,7 @@ namespace VWOSdk.Tests
             var mockSettingProcessor = Mock.GetSettingsProcessor();
             VWO.Configure(mockSettingProcessor.Object);
 
-            var vwoClient = VWO.CreateInstance(inValidSettings);
+            var vwoClient = VWO.Launch(inValidSettings);
             Assert.Null(vwoClient);
 
             mockValidator.Verify(mock => mock.SettingsFile(It.IsAny<Settings>()), Times.Once);
@@ -138,7 +138,7 @@ namespace VWOSdk.Tests
             //Mock.SetupProcessAndBucket(mockSettingProcessor, returnValue: null);
             VWO.Configure(mockSettingProcessor.Object);
 
-            var vwoClient = VWO.CreateInstance(inValidSettings);
+            var vwoClient = VWO.Launch(inValidSettings);
             Assert.Null(vwoClient);
 
             mockValidator.Verify(mock => mock.SettingsFile(It.IsAny<Settings>()), Times.Once);