+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.5.1] - 2020-04-13
+### Breaking Changes
+
+- `CreateInstance` API is renamed to `Launch` API.
+- `custom_variables` key inside options is renamed to `customVariables`
+- `revenue_value` key inside options is renamed to `revenueValue`
+
+### Changed
+
+- `GetVariation` can be used as an alternative for `GetVariationName` for backward compatibility.
+- `GetVariationName` API does not send any impression
+
 ## [1.5.0] - 2020-02-20
 ### Breaking Changes
 To prevent ordered arguments and increasing use-cases, we are moving all optional arguments to be passed via `options`.
@@ -27,7 +39,7 @@ bool isSuccessful = vwoClientInstance.Track(campaignKey, userId, goalIdentifier,
 Dictionary<string, dynamic> options = new Dictionary<string, dynamic>()
 {
     {
-        "custom_variable": new Dictionary<string, dynamic>()
+        "custom_variables": new Dictionary<string, dynamic>()
         {
             {"value", 10}
         }